Añadida funcionalidad de mostrar los alojamientos en el destino diario
Este commit aporta al issue #4 la funcionalidad de obtener los puntos de interés como alojamientos, restaurantes y demás en el destino diario.

+++ src/server/src/main/java/com/santiagoapp/routes/model/RoutePlaceBuilder.java
@@ -3,7 +3,7 @@ package com.santiagoapp.routes.model;
 public final class RoutePlaceBuilder {
     private String name;
     private double[] coordinates;
-    private int rating;
+    private double rating;
 
     private RoutePlaceBuilder() {
     }
@@ -22,7 +22,7 @@ public final class RoutePlaceBuilder {
         return this;
     }
 
-    public RoutePlaceBuilder withRating(int rating) {
+    public RoutePlaceBuilder withRating(double rating) {
         this.rating = rating;
         return this;
     }
